[step1] chore : 질문 삭제하기 기능 리팩토링
