--- src/main/java/qna/CannotDeleteException.java
@@ -1,9 +1,0 @@
-package qna;
-
-public class CannotDeleteException extends Exception {
-    private static final long serialVersionUID = 1L;
-
-    public CannotDeleteException(String message) {
-        super(message);
-    }
-}
