--- src/main/java/qna/ForbiddenException.java
@@ -1,10 +1,0 @@
-package qna;
-
-public class ForbiddenException extends RuntimeException{
-    public ForbiddenException() {
-    }
-
-    public ForbiddenException(String message) {
-        super(message);
-    }
-}
