--- src/main/java/qna/domain/ContentType.java
@@ -1,5 +1,0 @@
-package qna.domain;
-
-public enum ContentType {
-    QUESTION, ANSWER;
-}
