--- src/main/java/qna/domain/DeleteHistory.java
@@ -1,55 +1,0 @@
-package qna.domain;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.util.Objects;
-
-@Entity
-public class DeleteHistory {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private ContentType contentType;
-
-    private Long contentId;
-
-    @ManyToOne
-    @JoinColumn(foreignKey = @ForeignKey(name = "fk_deletehistory_to_user"))
-    private User deletedBy;
-
-    private LocalDateTime createDate = LocalDateTime.now();
-
-    public DeleteHistory() {
-    }
-
-    public DeleteHistory(ContentType contentType, Long contentId, User deletedBy, LocalDateTime createDate) {
-        this.contentType = contentType;
-        this.contentId = contentId;
-        this.deletedBy = deletedBy;
-        this.createDate = createDate;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        DeleteHistory that = (DeleteHistory) o;
-        return Objects.equals(id, that.id) &&
-                contentType == that.contentType &&
-                Objects.equals(contentId, that.contentId) &&
-                Objects.equals(deletedBy, that.deletedBy);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, contentType, contentId, deletedBy);
-    }
-
-    @Override
-    public String toString() {
-        return "DeleteHistory [id=" + id + ", contentType=" + contentType + ", contentId=" + contentId + ", deletedBy="
-                + deletedBy + ", createDate=" + createDate + "]";
-    }
-}
