+++ src/main/java/step1/qna/CannotDeleteException.java
@@ -1,0 +1,9 @@
+package step1.qna;
+
+public class CannotDeleteException extends Exception {
+	private static final long serialVersionUID = 1L;
+
+	public CannotDeleteException(String message) {
+		super(message);
+	}
+}
