+++ src/main/java/step1/qna/ForbiddenException.java
@@ -1,0 +1,10 @@
+package step1.qna;
+
+public class ForbiddenException extends RuntimeException {
+	public ForbiddenException() {
+	}
+
+	public ForbiddenException(String message) {
+		super(message);
+	}
+}
