+++ src/main/java/step1/qna/NotFoundException.java
@@ -1,4 +1,4 @@
-package qna;
+package step1.qna;
 
 public class NotFoundException extends RuntimeException {
 }
