+++ src/main/java/step1/qna/domain/AnswerRepository.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package step1.qna.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
-    List<Answer> findByQuestionAndDeletedFalse(Question question);
+	List<Answer> findByQuestionAndDeletedFalse(Question question);
 
-    Optional<Answer> findByIdAndDeletedFalse(Long id);
+	Optional<Answer> findByIdAndDeletedFalse(Long id);
 }
