+++ src/main/java/step1/qna/domain/ContentType.java
@@ -1,0 +1,5 @@
+package step1.qna.domain;
+
+public enum ContentType {
+	QUESTION, ANSWER
+}
