+++ src/main/java/step1/qna/domain/DeleteHistoryRepository.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package step1.qna.domain;
 
 import org.springframework.data.repository.CrudRepository;
