+++ src/main/java/step1/qna/domain/Question.java
@@ -1,0 +1,82 @@
+package step1.qna.domain;
+
+import org.hibernate.annotations.Where;
+import step1.qna.CannotDeleteException;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+public class Question extends AbstractEntity {
+	private static final String EXCEPTION_MSG = "질문을 삭제할 권한이 없습니다.";
+
+	@Column(length = 100, nullable = false)
+	private String title;
+
+	@Lob
+	private String contents;
+
+	@ManyToOne
+	@JoinColumn(foreignKey = @ForeignKey(name = "fk_question_writer"))
+	private User writer;
+
+	@OneToMany(mappedBy = "question", cascade = CascadeType.ALL)
+	@Where(clause = "deleted = false")
+	@OrderBy("id ASC")
+	private final List<Answer> answers = new ArrayList<>();
+
+	private boolean deleted = false;
+
+	public Question() {
+	}
+
+	public Question(String title, String contents) {
+		this.title = title;
+		this.contents = contents;
+	}
+
+	public Question(long id, String title, String contents) {
+		super(id);
+		this.title = title;
+		this.contents = contents;
+	}
+
+	public User getWriter() {
+		return writer;
+	}
+
+	public Question writeBy(User loginUser) {
+		this.writer = loginUser;
+		return this;
+	}
+
+	public void addAnswer(Answer answer) {
+		answer.toQuestion(this);
+		answers.add(answer);
+	}
+
+	public void checkDeleteCondition(User loginUser) throws CannotDeleteException {
+		if(!this.writer.equals(loginUser)) {
+			throw new CannotDeleteException(EXCEPTION_MSG);
+		}
+	}
+
+	public void deleteAndRecord(List<DeleteHistory> deleteHistories){
+		this.deleted = true;
+		deleteHistories.add(new DeleteHistory(ContentType.QUESTION, this));
+	}
+
+	public boolean isDeleted() {
+		return deleted;
+	}
+
+	public List<Answer> getAnswers() {
+		return answers;
+	}
+
+	@Override
+	public String toString() {
+		return "Question [id=" + getId() + ", title=" + title + ", contents=" + contents + ", writer=" + writer + "]";
+	}
+}
