+++ src/main/java/step1/qna/domain/QuestionRepository.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package step1.qna.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
-    List<Question> findByDeletedFalse();
+	List<Question> findByDeletedFalse();
 
-    Optional<Question> findByIdAndDeletedFalse(Long id);
+	Optional<Question> findByIdAndDeletedFalse(Long id);
 }
