+++ src/main/java/step1/qna/domain/UserRepository.java
@@ -1,9 +1,9 @@
-package qna.domain;
+package step1.qna.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUserId(String userId);
+	Optional<User> findByUserId(String userId);
 }
