+++ src/main/java/step1/qna/service/DeleteHistoryService.java
@@ -1,0 +1,26 @@
+package step1.qna.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+import step1.qna.domain.DeleteHistory;
+import step1.qna.domain.DeleteHistoryRepository;
+
+import javax.annotation.Resource;
+import java.util.List;
+
+@Service("deleteHistoryService")
+public class DeleteHistoryService {
+	@Resource(name = "deleteHistoryRepository")
+	private DeleteHistoryRepository deleteHistoryRepository;
+
+	@Transactional(propagation = Propagation.REQUIRES_NEW)
+	public void saveAll(List<DeleteHistory> deleteHistories) {
+		deleteHistoryRepository.saveAll(deleteHistories);
+	}
+
+	@Transactional(propagation = Propagation.REQUIRES_NEW)
+	public void save(DeleteHistory deleteHistory) {
+		deleteHistoryRepository.save(deleteHistory);
+	}
+}
