+++ src/main/java/step1/qna/service/QnAService.java
@@ -1,0 +1,48 @@
+package step1.qna.service;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import step1.qna.CannotDeleteException;
+import step1.qna.NotFoundException;
+import step1.qna.domain.*;
+
+import javax.annotation.Resource;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+@Service("qnaService")
+public class QnAService {
+	private static final Logger log = LoggerFactory.getLogger(QnAService.class);
+
+	@Resource(name = "questionRepository")
+	private QuestionRepository questionRepository;
+
+	@Resource(name = "answerRepository")
+	private AnswerRepository answerRepository;
+
+	@Resource(name = "deleteHistoryService")
+	private DeleteHistoryService deleteHistoryService;
+
+	@Transactional(readOnly = true)
+	public Question findQuestionById(Long id) {
+		return questionRepository.findByIdAndDeletedFalse(id)
+				.orElseThrow(NotFoundException::new);
+	}
+
+	@Transactional
+	public void deleteQuestion(User loginUser, long questionId) throws CannotDeleteException {
+		Question question = findQuestionById(questionId);
+		question.checkDeleteCondition(loginUser);
+
+		AnswerList answerList = new AnswerList(question);
+		answerList.checkDeleteCondition(loginUser);
+
+		List<DeleteHistory> deleteHistories = new ArrayList<>();
+		question.deleteAndRecord(deleteHistories);
+		answerList.deleteAndRecord(deleteHistories);
+		deleteHistoryService.saveAll(deleteHistories);
+	}
+}
