--- src/test/java/qna/domain/AnswerTest.java
@@ -1,6 +1,0 @@
-package qna.domain;
-
-public class AnswerTest {
-    public static final Answer A1 = new Answer(UserTest.JAVAJIGI, QuestionTest.Q1, "Answers Contents1");
-    public static final Answer A2 = new Answer(UserTest.SANJIGI, QuestionTest.Q1, "Answers Contents2");
-}
