+++ src/test/java/step1/qna/domain/UserTest.java
@@ -1,4 +1,6 @@
-package qna.domain;
+package step1.qna.domain;
+
+import step1.qna.domain.User;
 
 public class UserTest {
     public static final User JAVAJIGI = new User(1L, "javajigi", "password", "name", "javajigi@slipp.net");
